Rip out uom because it is not capable of the automation I expected
uom claimed that it normalized values to the base unit, and I thought they meant "find the least energy way to represent this unit".
example: if you only interact with milliamps, the base unit would be automatically found to be milliamps, and conversion code would happen at boundaries .
upon actually reading this however, the base unit is static. which is disappointing.